feat(resolver): Finish implements GraphQL

+++ backend/resolvers/Mutation/usuario.js
@@ -14,7 +14,8 @@ const mutations = {
     });
   },
 
-  async novoUsuario(_, { dados }) {
+  async novoUsuario(_, { dados }, ctx) {
+    ctx && ctx.validarAdmin();
     try {
       const idsPerfis = [];
 
@@ -45,37 +46,43 @@ const mutations = {
       throw new Error(e.sqlMessage);
     }
   },
-  async excluirUsuario(_, args) {
+  async excluirUsuario(_, args, ctx) {
+    ctx && ctx.validarAdmin();
     try {
       const usuario = await obterUsuario(_, args);
       if (usuario) {
         const { id } = usuario;
-        await db("usuarios_perfis").where({ usuario_id: id }).delete();
-        await db("usuarios").where({ id }).delete();
+        if (ctx.admin && dados.perfis) {
+          await db("usuarios_perfis").where({ usuario_id: id }).delete();
+          for (let filtro of dados.perfis) {
+            const perfil = await obterPerfil(_, { filtro });
+            if (perfil) {
+              await db("usuarios_perfis")
+                .where({ perfil_id: perfil.id, usuario_id: id })
+                .delete();
+            }
+          }
+        }
       }
       return usuario;
     } catch (e) {
       throw new Error(e.sqlMessage);
     }
   },
-  async alterarUsuario(_, { filtro, dados }) {
+  async alterarUsuario(_, { filtro, dados }, ctx) {
+    ctx && ctx.validarUsuarioFiltro(filtro);
     try {
       const usuario = await obterUsuario(_, { filtro });
       if (usuario) {
         const { id } = usuario;
-        if (dados.perfis) {
+        if (ctx.admin && dados.perfis) {
           await db("usuarios_perfis").where({ usuario_id: id }).delete();
-
           for (let filtro of dados.perfis) {
-            const perfil = await obterPerfil(_, {
-              filtro,
-            });
-
+            const perfil = await obterPerfil(_, { filtro });
             if (perfil) {
-              await db("usuarios_perfis").insert({
-                perfil_id: perfil.id,
-                usuario_id: id,
-              });
+              await db("usuarios_perfis")
+                .where({ perfil_id: perfil.id, usuario_id: id })
+                .delete();
             }
           }
         }
